% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{A file read function}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to the file containing data.}
}
\value{
A tibble with data.
}
\description{
\code{fars_read} function reads your data file downloaded from the
\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{Fatality Analysis Reporting System}.
If file you are trying to read doesn't exist on the fars database then you will get a message \code{filename does not exist}.
Any messages associated with downloading the file will be suppressed. Data will be read using the \code{read_csv} function from the \code{readr} package.
Dataframe will be stored in \code{tbl_df} format using the \code{dplyr} package
}
\examples{
\dontrun{
fars_complete_2013 <- fars_read("accident_2013-csv.bz2")
}
}
