% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read accident information by Month & Year for multiple FARS report files into a list of tibbles}
\usage{
fars_read_years(years)
}
\arguments{
\item{argument}{\code{years} takes list of years as an argument.}
}
\value{
A list of tibbles. Each tibble contains just two columns of the original
FARS data set (\code{MONTH} and \code{year}). Each row corresponds to a single accident.
}
\description{
\code{fars_read_years} function takes list of years as an argument, create file names, load data files,
mutate year and then select month and year.  If data for any of the years has been already downloaded then that data
is not imported.  Instead data from \code{tryCatch} is used
}
\details{
This function is not designed for direct user interaction.
It is a helper function for \code{fars_summarize_years}.
If any of the objects requested via input is not available as a year file
 or is not coercible to integer an "invalid year" error message returns.
}
\examples{
\dontrun{
fars_read_years(years = c(2013, 2014))
}
}
