% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize FARS report accidents by month and year in a tibble}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{\code{years}}{takes list of years as an argument.}
}
\value{
a tibble showing summary of number of accidents by month
  for each input year.
}
\description{
\code{fars_summarize_years} function takes list of years as an argument and
  provide the total number of accidents by month and year.
}
\examples{
\dontrun{
fars_summarize_years(years = c(2013, 2014))
}
}
