% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{Create file name using year}
\usage{
make_filename(year)
}
\arguments{
\item{argument}{\code{year} takes the year for which the data file is to be made.}
}
\value{
A character vector containing the complete file name of a compressed
  FARS report for a given year provided to the function as an argument.
}
\description{
\code{make_filename} function uses year as an argument and create file name by embedding \code{year} given as an argument to create a file.
File name is created by pasting \code{accident} and the input year and saving it in \code{csv.bz2} format.
}
\examples{
make_filename(2013)
}
